+++ Sources/MapboxMaps/Gestures/GestureHandlers/GestureHandler.swift
@@ -2,6 +2,10 @@ import UIKit
 
 internal protocol GestureHandlerDelegate: AnyObject {
     func gestureBegan(for gestureType: GestureType)
+
+    func gestureEnded(for gestureType: GestureType, willDecelerate: Bool)
+
+    func driftEnded(for gestureType: GestureType)
 }
 
 internal class GestureHandler: NSObject {